rename init_framework to hass_mock
