+++ appdaemontestframework/__init__.py
@@ -1,5 +1,5 @@
 from appdaemontestframework.assert_that import AssertThatWrapper
 from appdaemontestframework.given_that import GivenThatWrapper
 from appdaemontestframework.time_travel import TimeTravelWrapper
-from appdaemontestframework.init_framework import HassMock
+from appdaemontestframework.hass_mock import HassMock
 from appdaemontestframework.automation_fixture import automation_fixture
